fix: specify int32

+++ receiver/kafkareceiver/factory.go
@@ -42,11 +42,11 @@ const (
 	defaultAutoCommitInterval = 1 * time.Second
 
 	// default from sarama.NewConfig()
-	defaultMinFetchSize = 1
+	defaultMinFetchSize = int32(1)
 	// default from sarama.NewConfig()
-	defaultDefaultFetchSize = 1048576
+	defaultDefaultFetchSize = int32(1048576)
 	// default from sarama.NewConfig()
-	defaultMaxFetchSize = 0
+	defaultMaxFetchSize = int32(0)
 )
 
 var errUnrecognizedEncoding = fmt.Errorf("unrecognized encoding")
